requirements.txt
# Dependências para automação de testes IPTV e envio via WhatsApp
# Bibliotecas principais
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# Dependências para automação web
selenium==4.15.2
webdriver-manager==4.0.1

# Dependências para logging e manipulação de dados
logging==0.4.9.6
pyyaml==6.0.1

# Dependências para deploy no Render.com
Werkzeug==2.3.7
